Video Demos & Style Changes
Added video demos for projects. The Education and Sprite projects now have a video demo to view. These demo videos can be viewed via clicking the new button on those project elements.

Added a diagram showing the languages I know the most, also applied style changes to image to fit in better.

Made the canvas now not resize on mobile width and height changes. This was done so on mobile the layout didnt jump from the height resize of the canvas when first scrolling